pyproject.toml
[project]
name = "quickdemos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.11.1",
    "foundry-local-sdk>=0.5.1",
    "openai>=2.15.0",
]
